TGUI 0.6.9

+++ TGUI/src/TGUI/AnimatedPicture.cpp
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //
 // TGUI - Texus's Graphical User Interface
-// Copyright (C) 2012-2014 Bruno Van de Velde (vdv_b@tgui.eu)
+// Copyright (C) 2012-2015 Bruno Van de Velde (vdv_b@tgui.eu)
 //
 // This software is provided 'as-is', without any express or implied warranty.
 // In no event will the authors be held liable for any damages arising from the use of this software.
@@ -24,7 +24,7 @@
 
 
 #include <TGUI/AnimatedPicture.hpp>
-#include <TGUI\TGUI.hpp>
+#include <TGUI/TGUI.hpp>
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 namespace tgui
@@ -208,7 +208,7 @@ namespace tgui
         if (frame >= m_Textures.size())
         {
             // Display the last frame
-            m_CurrentFrame = static_cast<int>(m_Textures.size())-1;
+            m_CurrentFrame = m_Textures.size()-1;
             return false;
         }
 
@@ -239,7 +239,7 @@ namespace tgui
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-	size_t AnimatedPicture::getFrames() const
+    unsigned int AnimatedPicture::getFrames() const
     {
         return m_Textures.size();
     }
